Extracted `main` functionality into a standalone module and adjusted imports for cleaner structure.
